+++ sources/engine/sources/core/entities/StaticObject.cpp
@@ -6,7 +6,7 @@
 #include "level/Level.h"
 
 
-StaticObject::StaticObject(WireFrame* mesh, Texture* texture, int shader, const float posX, const float posY, int scale, float zBuffer) :
+StaticObject::StaticObject(WireFrame* mesh, Texture* texture, int shader, const float posX, const float posY, int scale, int zBuffer) :
 	GameObject(mesh, texture, shader, posX, posY, scale, INPUT_HANDLE_PROFILE_NONE)
 {
 	m_zbuffer = zBuffer;
